Reorganize and remove dependencies, part 1
