+++ app-backend/.sbtopts
@@ -1,3 +1,3 @@
--Dsbt.global.base=project/.sbtboot
 -Dsbt.boot.directory=project/.boot
+-Dsbt.global.base=project/.sbtboot
 -Dsbt.ivy.home=project/.ivy
