+++ app-backend/api/src/main/scala/config/Routes.scala
@@ -3,7 +3,7 @@ package com.rasterfoundry.api.config
 import akka.http.scaladsl.server.Route
 import cats.effect.IO
 import com.rasterfoundry.api.Codec._
-import com.rasterfoundry.authentication.Authentication
+import com.rasterfoundry.akkautil.{Authentication, CommonHandlers}
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import doobie.util.transactor.Transactor
 import io.circe._
@@ -14,7 +14,7 @@ import doobie.Fragments.in
 import doobie.postgres._
 import doobie.postgres.implicits._
 
-trait ConfigRoutes extends Authentication {
+trait ConfigRoutes extends CommonHandlers with Authentication {
   val xa: Transactor[IO]
   val configRoutes: Route = {
     pathEndOrSingleSlash {
