+++ app-backend/api/src/main/scala/exports/Routes.scala
@@ -8,7 +8,7 @@ import akka.http.scaladsl.server.Route
 import cats.data._
 import cats.effect.IO
 import cats.implicits._
-import com.rasterfoundry.authentication.Authentication
+import com.rasterfoundry.akkautil._
 import com.rasterfoundry.common._
 import com.rasterfoundry.database.ExportDao
 import com.rasterfoundry.database.filter.Filterables._
