+++ app-backend/api/src/main/scala/team/Routes.scala
@@ -1,7 +1,10 @@
 package com.rasterfoundry.api.team
 
-import com.rasterfoundry.authentication.Authentication
-import com.rasterfoundry.common.{CommonHandlers, UserErrorHandler}
+import com.rasterfoundry.akkautil.{
+  Authentication,
+  CommonHandlers,
+  UserErrorHandler
+}
 import com.rasterfoundry.database._
 import com.rasterfoundry.database.filter.Filterables._
 import com.rasterfoundry.datamodel._
