+++ app-backend/api/src/main/scala/tools/Routes.scala
@@ -1,6 +1,6 @@
 package com.rasterfoundry.api.tool
 
-import com.rasterfoundry.authentication.Authentication
+import com.rasterfoundry.akkautil._
 import com.rasterfoundry.common._
 import com.rasterfoundry.common.ast._
 import com.rasterfoundry.datamodel._
