--- app-backend/authentication/src/main/scala/package.scala
@@ -1,3 +1,0 @@
-package com.rasterfoundry.authentication
-
-package object authentication {}
