+++ app-backend/backsplash/src/main/resources/application.conf
@@ -1,0 +1,10 @@
+auth0 {
+  clientId = ""
+  clientId = ${?AUTH0_CLIENT_ID}
+
+  domain = ""
+  domain = ${?AUTH0_DOMAIN}
+
+  jwksURL = ""
+  jwksURL = ${?AUTH0_JWKS_URL}
+}
