+++ app-backend/backsplash/src/main/scala/io/Cog.scala
@@ -11,7 +11,6 @@ import geotrellis.proj4.{io => _, _}
 import geotrellis.raster.{CellSize, Raster, RasterExtent, io => _, _}
 import geotrellis.raster.io.geotiff.AutoHigherResolution
 import geotrellis.server.cog.util.CogUtils
-import geotrellis.spark.{io => _}
 import geotrellis.vector.{Extent, Polygon, Projected}
 
 object Cog extends RollbarNotifier {
