+++ app-backend/build.sbt
@@ -49,7 +49,6 @@ lazy val commonSettings = Seq(
     Resolver.bintrayRepo("azavea", "geotrellis"),
     Resolver.bintrayRepo("lonelyplanet", "maven"),
     Resolver.bintrayRepo("guizmaii", "maven"),
-    Resolver.bintrayRepo("kwark", "maven"), // Required for Slick 3.1.1.2, see https://github.com/azavea/raster-foundry/pull/1576
     "locationtech-releases" at "https://repo.locationtech.org/content/groups/releases",
     "locationtech-snapshots" at "https://repo.locationtech.org/content/groups/snapshots",
     Resolver.bintrayRepo("naftoligug", "maven"),
@@ -182,14 +181,6 @@ lazy val loggingDependencies = List(
   "ch.qos.logback" % "logback-classic" % "1.1.7"
 )
 
-lazy val slickDependencies = List(
-  Dependencies.slick,
-  Dependencies.slickPG,
-  Dependencies.slickPGSpray,
-  Dependencies.geotrellisSlick
-    .exclude("postgresql", "postgresql")
-)
-
 lazy val metricsDependencies = List(
   Dependencies.kamonCore,
   Dependencies.kamonStatsd,
@@ -224,24 +215,23 @@ lazy val apiDependencies = dbDependencies ++ migrationsDependencies ++
   Dependencies.akkastream,
   Dependencies.akkaSlf4j,
   Dependencies.akkaHttpExtensions,
+  Dependencies.awsStsSdk,
+  Dependencies.betterFiles,
   Dependencies.commonsIO,
-  Dependencies.ammoniteOps,
-  Dependencies.geotrellisSlick,
   Dependencies.geotrellisS3,
   Dependencies.geotrellisShapefile,
   Dependencies.betterFiles,
   Dependencies.caffeine,
   Dependencies.scaffeine,
   Dependencies.findbugAnnotations,
-  Dependencies.dropbox,
-  Dependencies.awsStsSdk
+  Dependencies.dropbox
 )
 
 lazy val root = Project("root", file("."))
   .settings(commonSettings: _*)
   .settings(noPublishSettings)
   .aggregate(api,
-             authentication,
+             akkautil,
              db,
              common,
              migrations,
@@ -253,10 +243,7 @@ lazy val root = Project("root", file("."))
              backsplash)
 
 lazy val api = Project("api", file("api"))
-  .dependsOn(db,
-             datamodel,
-             common % "test->test;compile->compile",
-             authentication)
+  .dependsOn(db, datamodel, common % "test->test;compile->compile", akkautil)
   .settings(apiSettings: _*)
   .settings(resolvers += Resolver.bintrayRepo("hseeberger", "maven"))
   .settings({
@@ -268,25 +255,16 @@ lazy val common = Project("common", file("common"))
   .settings(apiSettings: _*)
   .settings({
     libraryDependencies ++= testDependencies ++ Seq(
-      Dependencies.nimbusJose,
-      Dependencies.akka,
-      Dependencies.akkahttp,
-      Dependencies.akkaCirceJson,
       Dependencies.commonsIO,
       Dependencies.caffeine,
       Dependencies.scaffeine,
       Dependencies.elasticacheClient,
       Dependencies.geotrellisS3,
       Dependencies.findbugAnnotations,
-      Dependencies.ammoniteOps,
       Dependencies.chill,
       Dependencies.catsCore,
       Dependencies.awsBatchSdk,
-      Dependencies.awsStsSdk,
       Dependencies.rollbar,
-      Dependencies.doobiePostgres,
-      Dependencies.doobiePostgresCirce,
-      Dependencies.geotrellisSlick.exclude("postgresql", "postgresql"),
       Dependencies.apacheCommonsEmail
     )
   })
@@ -307,7 +285,6 @@ lazy val db = Project("db", file("db"))
       "net.postgis" % "postgis-jdbc-jtsparser" % "2.2.1",
       "org.locationtech.jts" % "jts-core" % "1.15.0",
       "com.lonelyplanet" %% "akka-http-extensions" % "0.4.15",
-      Dependencies.geotrellisSlick.exclude("postgresql", "postgresql")
     )
   })
 
@@ -324,16 +301,12 @@ lazy val datamodel = Project("datamodel", file("datamodel"))
   .settings(resolvers += Resolver.bintrayRepo("azavea", "geotrellis"))
   .settings({
     libraryDependencies ++= loggingDependencies ++ Seq(
-      Dependencies.geotrellisSlick % "provided",
       Dependencies.geotrellisVectorTestkit,
       Dependencies.geotrellisRaster,
       Dependencies.geotrellisGeotools,
       Dependencies.geotools,
       Dependencies.circeCore,
       Dependencies.circeGenericExtras,
-      Dependencies.akka,
-      Dependencies.akkahttp,
-      Dependencies.betterFiles,
       Dependencies.scalaCheck,
       Dependencies.circeTest,
       "com.lonelyplanet" %% "akka-http-extensions" % "0.4.15" % "test",
@@ -352,17 +325,8 @@ lazy val batch = Project("batch", file("batch"))
       Dependencies.geotrellisS3,
       Dependencies.geotrellisUtil,
       Dependencies.geotrellisRaster,
-      Dependencies.akka,
-      Dependencies.akkahttp,
-      Dependencies.akkaHttpCors,
-      Dependencies.akkaCirceJson,
-      Dependencies.akkastream,
-      Dependencies.akkaSlf4j,
-      Dependencies.akkaSprayJson,
-      Dependencies.geotrellisSlick,
       Dependencies.sparkCore,
       Dependencies.hadoopAws,
-      Dependencies.awsSdk,
       Dependencies.scopt,
       Dependencies.ficus,
       Dependencies.dnsJava,
@@ -401,7 +365,7 @@ import _root_.io.gatling.sbt.GatlingPlugin
 lazy val tile = Project("tile", file("tile"))
   .dependsOn(datamodel,
              common % "test->test;compile->compile",
-             authentication,
+             akkautil,
              geotrellis)
   .dependsOn(tool)
   .enablePlugins(GatlingPlugin)
@@ -413,7 +377,6 @@ lazy val tile = Project("tile", file("tile"))
       Dependencies.spark,
       Dependencies.geotrellisSpark,
       Dependencies.geotrellisS3,
-      Dependencies.akkaSprayJson,
       Dependencies.akkaCirceJson,
       Dependencies.akkaHttpCors,
       Dependencies.akkastream,
@@ -449,15 +412,13 @@ lazy val tool = Project("tool", file("tool"))
       Dependencies.sparkCore,
       Dependencies.geotrellisSpark,
       Dependencies.geotrellisRaster,
-      Dependencies.geotrellisRasterTestkit,
-      Dependencies.shapeless,
+      Dependencies.geotrellisRasterTestkit % "test",
       Dependencies.scalatest,
       Dependencies.circeCore,
       Dependencies.circeGeneric,
       Dependencies.circeParser,
       Dependencies.circeOptics,
       Dependencies.scalaCheck,
-      Dependencies.scalaz,
       Dependencies.mamlJvm
     )
   })
@@ -474,7 +435,7 @@ lazy val geotrellis = Project("geotrellis", file("geotrellis"))
     )
   })
 
-lazy val authentication = Project("authentication", file("authentication"))
+lazy val akkautil = Project("akkautil", file("akkautil"))
   .dependsOn(common, db)
   .settings(commonSettings: _*)
   .settings({
@@ -500,7 +461,7 @@ lazy val bridge = Project("bridge", file("bridge"))
 
 // maml / better-abstracted tile server
 lazy val backsplash = Project("backsplash", file("backsplash"))
-  .dependsOn(authentication, geotrellis, db, tool)
+  .dependsOn(geotrellis, db, tool)
   .settings(commonSettings: _*)
   .settings(noPublishSettings)
   .settings(fork in run := true)
@@ -515,7 +476,8 @@ lazy val backsplash = Project("backsplash", file("backsplash"))
       Dependencies.http4sCirce,
       Dependencies.http4sDSL,
       Dependencies.http4sServer,
-      Dependencies.mamlJvm
+      Dependencies.mamlJvm,
+      Dependencies.nimbusJose
     )
   })
   .settings(addCompilerPlugin("org.spire-math" %% "kind-projector" % "0.9.7"))
