+++ app-backend/common/src/main/scala/MalformedContentRejectionHandler.scala
@@ -1,4 +1,4 @@
-package com.rasterfoundry.common
+package com.rasterfoundry.akkautil
 
 import akka.http.scaladsl.server._
 import akka.http.scaladsl.model._
