+++ app-backend/common/src/main/scala/package.scala
@@ -1,0 +1,3 @@
+package com.rasterfoundry.akkautil
+
+package object akkautil {}
