--- app-backend/datamodel/src/main/scala/QueryResults.scala
@@ -1,5 +1,0 @@
-package com.rasterfoundry.datamodel
-
-import slick.dbio.DBIO
-
-final case class ListQueryResult[T](records: DBIO[Seq[T]], nRecords: DBIO[Int])
