+++ app-backend/project/Version.scala
@@ -5,10 +5,8 @@ object Version {
   val akkaHttpCors = "0.2.2"
   val akkaHttpExtensions = "0.4.13"
   val akkaSlf4j = "2.4.11"
-  val ammoniteOps = "0.7.0"
   val auth0 = "1.5.0"
   val awsBatchSdk = "1.11.196"
-  val awsSdk = "1.11.92"
   val awsStsSdk = "1.11.232"
   val betterFiles = "3.4.0"
   val caffeine = "2.3.5"
@@ -46,14 +44,9 @@ object Version {
   val scalaLogging = "3.9.0"
   val scalaTest = "3.0.1"
   val scalajHttp = "2.3.0"
-  val scalaz = "7.2.8"
   val scapegoat = "1.3.7"
   val scopt = "3.5.0"
-  val shapeless = "2.3.2"
-  val slf4j = "1.6.4"
-  val slick = "3.2.1"
   val slickMigrationAPI = "0.4.0"
-  val slickPG = "0.15.4"
   val spark = "2.2.0"
   val apacheCommonsEmail = "1.5"
 }
