+++ app-backend/tool/src/main/scala/RenderDefinition.scala
@@ -7,7 +7,6 @@ import io.circe._
 import io.circe.syntax._
 import io.circe.generic.semiauto._
 import io.circe.parser.decode
-import org.scalatest._
 import geotrellis.raster.render._
 
 final case class RenderDefinition(
