+++ app-backend/tool/src/main/scala/package.scala
@@ -5,7 +5,6 @@ import io.circe._
 import io.circe.syntax._
 import io.circe.generic.semiauto._
 import io.circe.parser.decode
-import org.scalatest._
 import geotrellis.raster.render._
 
 import scala.util.Try
