+++ scripts/bootstrap
@@ -32,7 +32,7 @@ function build_containers() {
         build
     docker-compose \
         -f "${DIR}/docker-compose.yml" \
-        pull postgres memcached   
+        pull postgres memcached
 
     popd
 }
@@ -52,7 +52,7 @@ function pull_sbtopts() {
 
     aws s3 cp "s3://${RF_SETTINGS_BUCKET}/.sbtopts" - >> app-backend/.sbtopts
     sort app-backend/.sbtopts | uniq > cleaned_sbtopts
-    mv cleaned_sbtopts app-backend/.sbtopts     
+    mv cleaned_sbtopts app-backend/.sbtopts
 
     popd
 }
